requirements.txt
python-docx
pypdf
Pillow
PyExifTool
pptx
openpyxl
customtkinter
